backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.5.0
pydantic==2.4.2
pydantic-settings==2.0.3
pyyaml==6.0.1
python-multipart==0.0.6
openai>=1.0.0
groq>=0.4.0
anthropic>=0.8.0
tenacity>=8.2.0
